RLP-796: rename transport entities (#122)
- rename TransportLayer class to TransportNode
- rename NodeTransport class to TransportLayer
