+++ raiden/network/transport/matrix/__init__.py
@@ -1,4 +1,5 @@
-from raiden.network.transport.matrix.transport import MatrixTransport, _RetryQueue, MatrixLightClientTransport, NodeTransport  # noqa
+from raiden.network.transport.matrix.transport import MatrixTransport, _RetryQueue, MatrixLightClientTransport  # noqa
+from transport.layer import Layer as TransportLayer  # noqa
 from raiden.network.transport.matrix.utils import (  # noqa
     AddressReachability,
     UserPresence,
