+++ transport/layer.py
@@ -1,64 +1,16 @@
-from abc import ABC, abstractmethod
-from typing import Any
+from typing import List
 
-from raiden.message_handler import MessageHandler
-from raiden.messages import Message
-from raiden.raiden_service import RaidenService
-from raiden.transfer.identifiers import QueueIdentifier
-from raiden.utils import Address
+from transport.node import Node as TransportNode
 
 
-class TransportLayer(ABC):
+class Layer:
 
-    def __init__(self, address: Address):
-        self._address = address  # source for messages transmitted over this layer
+    def __init__(self, hub_transport: TransportNode, light_client_transports: List[TransportNode]):
+        self.hub_transport = hub_transport
+        self.light_client_transports = light_client_transports
 
-    @property
-    def address(self):
-        return self._address
+    def add_light_client_transport(self, light_client_transport: TransportNode):
+        self.light_client_transports.append(light_client_transport)
 
-    @abstractmethod
-    def start(self, raiden_service: RaidenService, message_handler: MessageHandler, prev_auth_data: str):
-        """
-        Start the transport layer.
-        """
-
-    @abstractmethod
-    def stop(self):
-        """
-        Stop the transport layer.
-        """
-
-    @abstractmethod
-    def send_async(self, queue_identifier: QueueIdentifier, message: Message):
-        """
-        Queue the message for sending to recipient in the queue_identifier.
-        It may be called before transport is started, to initialize message queues.
-        The actual sending will be started only when the transport is started.
-        """
-
-    @abstractmethod
-    def start_health_check(self, address: Address):
-        """
-        Start health-check (status monitoring) for a peer.
-        It also whitelists the address to answer invites and listen for messages.
-        """
-
-    @abstractmethod
-    def whitelist(self, address: Address):
-        """
-        Whitelist peer address from which to receive communications.
-        This may be called before transport is started, to ensure events generated during start are handled properly.
-        """
-
-    @abstractmethod
-    def link_exception(self, callback: Any):
-        """
-        Add a callback function to be executed once the transport layer is halted due to an exception.
-        """
-
-    @abstractmethod
-    def join(self, timeout=None):
-        """
-        Wait until the transport layer finishes its pending tasks or the given timeout passes.
-        """
+    def remove_light_client_transport(self, light_client_transport: TransportNode):
+        self.light_client_transports.remove(light_client_transport)
